Create domain models
First step for creating a layered architecture for connector
https://github.com/interledger/five-bells-ledger/issues/112
